Move view related components to _partials
Moved them to _partials to be more in-sync with interactive design setup
