+++ src/views/teams/ViewTeamsIndex.vue
@@ -11,13 +11,14 @@
 </template>
 
 <script setup lang="ts">
-import LayoutSettings from '@/layouts/LayoutSettings.vue'
-import TeamsAdministratorRole from '@/components/Teams/TeamsAdministratorRole.vue'
-import TeamsHeader from '@/components/Teams/TeamsHeader.vue'
 import { useTeamsStore } from '@/stores/teams'
-import TeamsModalNewTeam from '@/components/Teams/TeamsModalNewTeam.vue'
-import TeamsMemberRole from '@/components/Teams/TeamsMemberRole.vue'
-import TeamsModalLeaveTeam from '@/components/Teams/TeamsModalLeaveTeam.vue'
+import LayoutSettings from '@/layouts/LayoutSettings.vue'
+
+import TeamsAdministratorRole from '@/views/teams/_partials/TeamsAdministratorRole.vue'
+import TeamsHeader from '@/views/teams/_partials/TeamsHeader.vue'
+import TeamsMemberRole from '@/views/teams/_partials/TeamsMemberRole.vue'
+import TeamsModalLeaveTeam from '@/views/teams/_partials/TeamsModalLeaveTeam.vue'
+import TeamsModalNewTeam from '@/views/teams/_partials/TeamsModalNewTeam.vue'
 
 const teamsStore = useTeamsStore()
 teamsStore.getTeams()
